fix: address third round of PR review feedback
- Use model.baseUrl in callGoogle when available, fall back to public endpoint (thread 1)
- Emit model-router:suggested event and log in suggest mode (thread 2)
- Extract shared fuzzy-match logic into src/match.ts, deduplicate classifier and resolver (thread 3)
- Simplify dead target.thinking guard to always apply thinking level on switch (thread 4)
